@@ -1,29 +1,71 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Microsoft.OpenApi.Models;
+﻿using Microsoft.EntityFrameworkCore;
+
+using Subrom.Services;
+using Subrom.Services.Interfaces;
+using Subrom.SubromAPI.Data;
 
 var builder = WebApplication.CreateBuilder(args);
 
-// Add services to the container.
+// Database
+var dbPath = Path.Combine(
+	Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData),
+	"Subrom",
+	"subrom.db"
+);
+Directory.CreateDirectory(Path.GetDirectoryName(dbPath)!);
 
+builder.Services.AddDbContext<SubromDbContext>(options =>
+	options.UseSqlite($"Data Source={dbPath}"));
+
+// Services
+builder.Services.AddScoped<IHashService, HashService>();
+
+// Controllers
 builder.Services.AddControllers();
-builder.Services.AddSwaggerGen(c => c.SwaggerDoc("v1", new() { Title = "Subrom", Version = "v1" }));
+
+// OpenAPI/Swagger
+builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSwaggerGen(options => {
+	options.SwaggerDoc("v1", new() {
+		Title = "Subrom API",
+		Version = "v1",
+		Description = "ROM management and verification API",
+	});
+});
+
+// CORS for React dev server
+builder.Services.AddCors(options => {
+	options.AddDefaultPolicy(policy => {
+		policy.WithOrigins("http://localhost:3000")
+			.AllowAnyHeader()
+			.AllowAnyMethod();
+	});
+});
 
 var app = builder.Build();
 
-// Configure the HTTP request pipeline.
-if (builder.Environment.IsDevelopment())
-{
+// Apply migrations on startup
+using (var scope = app.Services.CreateScope()) {
+	var db = scope.ServiceProvider.GetRequiredService<SubromDbContext>();
+	await db.Database.EnsureCreatedAsync();
+}
+
+// Configure pipeline
+if (app.Environment.IsDevelopment()) {
 	app.UseDeveloperExceptionPage();
 	app.UseSwagger();
-	app.UseSwaggerUI(c => c.SwaggerEndpoint("/swagger/v1/swagger.json", "Subrom v1"));
+	app.UseSwaggerUI(options => {
+		options.SwaggerEndpoint("/swagger/v1/swagger.json", "Subrom API v1");
+		options.RoutePrefix = "swagger";
+	});
 }
 
 app.UseHttpsRedirection();
-
+app.UseCors();
 app.UseAuthorization();
-
 app.MapControllers();
 
-app.Run();
+// Health check endpoint
+app.MapGet("/health", () => Results.Ok(new { status = "healthy", timestamp = DateTime.UtcNow }));
+
+await app.RunAsync();
